# Added vec based imresize using numpy leading to massive performance boost