I added an implementation of `imresizemex` called `imresizevec`using some reshaping of the `weights` array to support vec-multiplication with  `inimg` without needing any for-loops. To support the original implementation and the new one I added the `mode` flag that allows to choose either `"vec"` or `"org"`.  I kept everything else. I verified the implementation using the provided test case and get.

> Python/MatLab uint8 diff: min = 0 max = 0
> Python/Matlab double diff: min = 0 max = 0
> [Finished in 1.4s]

Benchmarking the new implementation for up scaling reveals

```python
img = np.random.randn(250,400,3)
%timeit -n 10 imresize(img, scalar_scale=2, mode="org")
%timeit -n 10 imresize(img, scalar_scale=2, mode="vec")
```
> 10 loops, best of 3: 6.47 s per loop
> 10 loops, best of 3: 68.1 ms per loop

and for downscaling

```python
img = np.random.randn(250,400,3)
%timeit -n 10 imresize(img, scalar_scale=0.5, mode="org")
%timeit -n 10 imresize(img, scalar_scale=0.5, mode="vec")
```
> 10 loops, best of 3: 884 ms per loop
> 10 loops, best of 3: 14.3 ms per loop

Hence, the vectorized implementation is for the given scenario up to **61-95** times faster.
Would be nice to have it in the repository to get closer to a useable imresize function to augment the data during training of CNNs for Super-resolution.