@fatheral 2018-09-17
Thank you, @2006pmach ! Now it really works very fast :)